requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt
#
blurhash==1.1.4
    # via mastodon-py
certifi==2023.7.22
    # via requests
charset-normalizer==3.2.0
    # via requests
decorator==5.1.1
    # via mastodon-py
idna==3.4
    # via requests
markdown-it-py==3.0.0
    # via rich
mastodon-py==1.8.1
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
pygments==2.15.1
    # via rich
python-dateutil==2.8.2
    # via mastodon-py
python-dotenv==1.0.0
    # via -r requirements.in
python-magic-bin==0.4.14
    # via mastodon-py
requests==2.31.0
    # via mastodon-py
rich==13.5.2
    # via -r requirements.in
six==1.16.0
    # via
    #   mastodon-py
    #   python-dateutil
urllib3==2.0.4
    # via requests
